ingest/requirements.txt
edpy==0.1.1
edapi
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
urllib3==2.0.7
schedule==1.2.1
click==8.1.7
bcrypt==4.1.2
